fix: tests and types
