+++ packages/parser/src/get-module-specifier-value.ts
@@ -1,0 +1,9 @@
+import type { ImportDeclaration } from 'ts-morph'
+
+export const getModuleSpecifierValue = (node: ImportDeclaration) => {
+  try {
+    return node.getModuleSpecifierValue()
+  } catch {
+    return
+  }
+}
